+++ lib/td_dd_web/controllers/search_controller.ex
@@ -1,9 +1,10 @@
 defmodule TdDdWeb.SearchController do
-  use TdDdWeb, :controller
-  import Canada, only: [can?: 2]
   use PhoenixSwagger
+  use TdDdWeb, :controller
+
+  import Canada, only: [can?: 2]
+
   alias TdDd.DataStructures.DataStructure
-  alias TdDdWeb.ErrorView
 
   @index_worker Application.get_env(:td_dd, :index_worker)
 
@@ -21,17 +22,8 @@ defmodule TdDdWeb.SearchController do
       @index_worker.reindex(:all)
       send_resp(conn, :accepted, "")
     else
-      false ->
-        conn
-        |> put_status(:forbidden)
-        |> put_view(ErrorView)
-        |> render("403.json")
-
-      _error ->
-        conn
-        |> put_status(:internal_server_error)
-        |> put_view(ErrorView)
-        |> render("500.json")
+      false -> render_error(conn, :forbidden)
+      _error -> render_error(conn, :internal_server_error)
     end
   end
 end