+++ priv/repo/migrations/20190917105053_update_data_structure_version_hashes.exs
@@ -1,0 +1,7 @@
+defmodule TdDd.Repo.Migrations.UpdateDataStructureVersionHashes do
+  use Ecto.Migration
+
+  def change do
+    execute("update data_structure_versions set hash = null, lhash = null, ghash = null;")
+  end
+end